Revert "not needed"
This reverts commit d3cf2883a6338e61f2e482e4cc0eec75aa4ba58a.

In attempt to try and figure out which changes are needed and which aren't, I
simply deleted them and rebuild it. It appears that that isn't enough, even
with the deletion of the whole layout overlay folder, the phone app kept working
in the emulator.

So to be conclusive: these five filles are the max. files that are needed to get
stop com.android.phone from FCing on ldpi devices. But it could be that there
should be less. I'll try to figure it out when I have the time.