Revert "Fix the size of the notification icons in the expanded status bar"
This reverts commit 99e9fdb31e84e5272e2e4049f5560f72f2982aa8.

It appeared that the icons are changed, but not as I was hoping they would.
As the avatars in the expanded status bar from Gmail notifications were
displayed properly, it should not be the "general" icon size that should be
changed, but only the icons/avatars from the Mms application.