Removed the 3G tweaks
They are probably the cause of rebooting when the mobile data connection
is enabled. The only way to get out of this rebooting cycle is either
disable the data connection before the device makes an attempt to connect
to the internet, or by reflashing the radio (which has to be done every
time the devices reboots).
Hopefully this commit fixes that issue.