Add some performance settings.
It seems that setting some of the performance settings through the
UI don't seem to have any effect. By adding them to the build.prop,
these settings are setted on every (re)boot.

Especially hard-setting the 16-bits transparrancy and the surface
dithering help in preventing some visual artefacts.

Thanks to XDA-member Justice™ for noticing and testing