Revert "Use ldpi-common repository for overlay files"
This reverts commit 8828c91109f3c0ad38fe48dadd9ecb5719a3bf4b.

The overlay files from Rallapag are quite different from the onces
from us. This can be seen in the pattern unlock screen for example,
and from the icons in the recent apps. Also at the home screen, if
you put two icons in one docking spot at the same hight of the app-
drawer button, you can see that they are not full visible on the app
screen.

So there are two options:
 - or we step away from the ldpi-common repository and use are own
   overlay files. This is what this commit does.
 - we use the ldpi-common repository, but with or own overlay files.
   There will be a commit for that soon.

So after both commits, we will use the own overlay files, but we can
switch easily to the ldpi-common of it is ought to be a better way to
work in (and if the changes are excepted).