Revert "remove: LibOmxCore (useless) and build libaudio from sources to support radio"
This reverts commit 5614542274395cddfba34135c129684627a80152.
Removal of both does succeed in a build-compable resource, but in which the speaker
("recording audio") doesn't work. I tried it in phone calls and in Google Voice from
the market. Furthermore, when I try to record a song with SoundHound, it states that
it doesn't find a working microphone. No clear errors in the logcat, but it seems that
there are problems in the copereration between libaudio and various other libs, including
libutils, libc and libaudioflinger, resuling in:
    W/AudioSystem(  307): AudioFlinger server died!
    W/AudioSystem(  307): AudioPolicyService server died!
    W/AudioSystem( 5533): AudioFlinger server died!
    W/AudioSystem( 5533): AudioPolicyService server died!

Reverting the part in which libaudio is build from source to see if that revert would
fix it.