Trebuchet: port wallpaper hack
Normally the wallpaper is drawn by the system in a separate window.
This means that the wallpaper can often be drawn out of sync with
the launcher. This hack gets the wallpaper bitmap and draws it
manually (in sync with the rest of the launcher). This increases
performance greatly.

This is a port from the cm-10.1 branch
http://review.cyanogenmod.org/#/c/27326/3